brownie_contract_dev/requirements.txt
eth-brownie>=1.10.0,<2.0.0
ipdb
